US-512975 Fixed unit tests
