+++ terratest/src/test/pega/data/values_backend_config_gke.yaml
@@ -1,0 +1,11 @@
+---
+global:
+  tier:
+    - name: "web"
+      ingress:
+        backendConfig:
+          connectionDraining:
+            drainingTimeoutSec: 100
+      service:
+        port: 80
+        targetPort: 8080
