--- terratest/src/test/pega/data/values_gke_backend_config.yaml
@@ -1,19 +1,0 @@
----
-global:
-  tier:
-    - name: "web"
-      nodeType: "WebUser"
-      requestor:
-        passivationTimeSec: 900
-      replicas: 1
-      deploymentStrategy:
-        rollingUpdate:
-          maxSurge: 1
-          maxUnavailable: 0
-        type: RollingUpdate
-      livenessProbe:
-        port: 8081
-      # Values for test - web
-      backendConfig:
-        connectionDraining:
-          drainingTimeoutSec: 80
